Refactor: Update import paths for hooks and services in components
- Changed import paths for useColorScheme in Collapsible and ParallaxScrollView components to reflect new directory structure.
- Updated UserListItem to import User from the correct services path, enhancing code organization and clarity.
