+++ shared/components/Collapsible.tsx
@@ -5,7 +5,7 @@ import { ThemedText } from '@/shared/components/ThemedText';
 import { ThemedView } from '@/shared/components/ThemedView';
 import { IconSymbol } from '@/shared/components/ui/IconSymbol';
 import { Colors } from '@/shared/constants/Colors';
-import { useColorScheme } from '@/hooks/useColorScheme';
+import { useColorScheme } from '@/shared/hooks/theme/useColorScheme';
 
 export function Collapsible({ children, title }: PropsWithChildren & { title: string }) {
   const [isOpen, setIsOpen] = useState(false);
